libflux: split reactor out to its own module
