+++ src/common/libflux/handle.c
@@ -32,7 +32,6 @@
 
 #include "handle_impl.h"
 #include "message.h"
-#include "reactor.h"
 #include "tagpool.h"
 
 #include "src/common/libutil/log.h"
@@ -40,18 +39,13 @@
 #include "src/common/libutil/xzmalloc.h"
 
 
-typedef struct reactor_struct *reactor_t;
-
-static void reactor_destroy (reactor_t r);
-static reactor_t reactor_create (void);
-
 struct flux_handle_struct {
     const struct flux_handle_ops *ops;
     int             flags;
     void            *impl;
-    reactor_t       reactor;
     zhash_t         *aux;
     tagpool_t       tagpool;
+    reactor_t       reactor;
 };
 
 flux_t flux_handle_create (void *impl, const struct flux_handle_ops *ops, int flags)
@@ -61,10 +55,10 @@ flux_t flux_handle_create (void *impl, const struct flux_handle_ops *ops, int fl
     h->flags = flags;
     if (!(h->aux = zhash_new()))
         oom ();
-    h->reactor = reactor_create ();
     h->ops = ops;
     h->impl = impl;
     h->tagpool = tagpool_create ();
+    h->reactor = flux_reactor_create (impl, ops);
     return h;
 }
 
@@ -73,11 +67,11 @@ void flux_handle_destroy (flux_t *hp)
     flux_t h;
 
     if (hp && (h = *hp)) {
+        zhash_destroy (&h->aux);
         if (h->ops->impl_destroy)
             h->ops->impl_destroy (h->impl);
-        zhash_destroy (&h->aux);
-        reactor_destroy (h->reactor);
         tagpool_destroy (h->tagpool);
+        flux_reactor_destroy (h->reactor);
         free (h);
         *hp = NULL;
     }
@@ -91,6 +85,11 @@ void flux_flags_set (flux_t h, int flags)
 void flux_flags_unset (flux_t h, int flags)
 {
     h->flags &= ~flags;
+}
+
+int flux_flags_get (flux_t h)
+{
+    return h->flags;
 }
 
 void *flux_aux_get (flux_t h, const char *name)
@@ -256,370 +255,9 @@ zctx_t *flux_get_zctx (flux_t h)
     return h->ops->get_zctx (h->impl);
 }
 
-/**
- ** Reactor
- **/
-
-typedef enum {
-    DSP_TYPE_MSG, DSP_TYPE_FD, DSP_TYPE_ZS, DSP_TYPE_TMOUT
-} dispatch_type_t;
-
-typedef struct {
-    dispatch_type_t type;
-    union {
-        struct {
-            flux_match_t match;
-            FluxMsgHandler fn;
-            void *arg;
-        } msg;
-        struct {
-            int fd;
-            short events;
-            FluxFdHandler fn;
-            void *arg;
-        } fd;
-        struct {
-            void *zs;
-            short events;
-            FluxZsHandler fn;
-            void *arg;
-        } zs;
-        struct {
-            int timer_id;
-            bool oneshot;
-            FluxTmoutHandler fn;
-            void *arg;
-        } tmout;
-    };
-} dispatch_t;
-
-struct reactor_struct {
-    zlist_t *dsp;
-    bool timeout_set;
-};
-
-static bool reactor_empty (reactor_t r)
+reactor_t flux_get_reactor (flux_t h)
 {
-    return ((zlist_size (r->dsp) == 0) && !r->timeout_set);
-}
-
-static dispatch_t *dispatch_create (dispatch_type_t type)
-{
-    dispatch_t *d = xzmalloc (sizeof (*d));
-    d->type = type;
-    return d;
-}
-
-static void dispatch_destroy (dispatch_t *d)
-{
-    if (d && d->type == DSP_TYPE_MSG && d->msg.match.topic_glob)
-        free (d->msg.match.topic_glob);
-    free (d);
-}
-
-static reactor_t reactor_create (void)
-{
-    reactor_t r = xzmalloc (sizeof (*r));
-    if (!(r->dsp = zlist_new ()))
-        oom ();
-    return r;
-}
-
-static void reactor_destroy (reactor_t r)
-{
-    dispatch_t *d;
-
-    while ((d = zlist_pop (r->dsp)))
-        dispatch_destroy (d);
-    zlist_destroy (&r->dsp);
-    free (r);
-}
-
-int flux_handle_event_msg (flux_t h, zmsg_t **zmsg)
-{
-    dispatch_t *d;
-    int type = 0;
-    int rc = 0;
-    bool match = false;
-
-    assert (zmsg != NULL);
-    assert (*zmsg != NULL);
-
-    if (flux_msg_get_type (*zmsg, &type) < 0)
-        goto done;
-    d = zlist_first (h->reactor->dsp);
-    while (d) {
-        if (d->type == DSP_TYPE_MSG && flux_msg_cmp (*zmsg, d->msg.match)) {
-            if (d->msg.fn)
-                rc = d->msg.fn (h, type, zmsg, d->msg.arg);
-            match = true;
-            break;
-        }
-        d = zlist_next (h->reactor->dsp);
-    }
-done:
-    if (h->flags & FLUX_FLAGS_TRACE) {
-        if (!match && *zmsg) {
-            char *topic = NULL;
-            (void)flux_msg_get_topic (*zmsg, &topic);
-            fprintf (stderr, "nomatch: %s '%s'\n", flux_msgtype_string (type),
-                     topic ? topic : "");
-        }
-    }
-    return rc;
-}
-
-int flux_handle_event_fd (flux_t h, int fd, short events)
-{
-    dispatch_t *d;
-    int rc = 0;
-
-    d = zlist_first (h->reactor->dsp);
-    while (d) {
-        if (d->type == DSP_TYPE_FD && d->fd.fd == fd
-                            && (d->fd.events & events) && d->fd.fn != NULL) {
-            rc = d->fd.fn (h, fd, events, d->fd.arg);
-            break;
-        }
-        d = zlist_next (h->reactor->dsp);
-    }
-    return rc;
-}
-
-int flux_handle_event_zs (flux_t h, void *zs, short events)
-{
-    dispatch_t *d;
-    int rc = 0;
-
-    d = zlist_first (h->reactor->dsp);
-    while (d) {
-        if (d->type == DSP_TYPE_ZS && d->zs.zs == zs
-                                   && (d->zs.events & events)) {
-            rc = d->zs.fn (h, zs, events, d->zs.arg);
-            break;
-        }
-        d = zlist_next (h->reactor->dsp);
-    }
-    return rc;
-}
-
-int flux_handle_event_tmout (flux_t h, int timer_id)
-{
-    dispatch_t *d;
-    int rc = 0;
-
-    d = zlist_first (h->reactor->dsp);
-    while (d) {
-        if (d->type == DSP_TYPE_TMOUT && d->tmout.fn != NULL
-                                      && d->tmout.timer_id == timer_id) {
-            rc = d->tmout.fn (h, d->tmout.arg);
-            break;
-        }
-        d = zlist_next (h->reactor->dsp);
-    }
-    if (d && d->tmout.oneshot)
-        flux_tmouthandler_remove (h, d->tmout.timer_id);
-    return rc;
-}
-
-int flux_msghandler_add (flux_t h, int typemask, const char *pattern,
-                         FluxMsgHandler cb, void *arg)
-{
-    dispatch_t *d;
-    int rc = -1;
-
-    if (typemask == 0 || !pattern || !cb) {
-        errno = EINVAL;
-        goto done;
-    }
-    d = dispatch_create (DSP_TYPE_MSG);
-    d->msg.match.typemask = typemask;
-    d->msg.match.matchtag = FLUX_MATCHTAG_NONE;
-    d->msg.match.bsize = 0;
-    d->msg.match.topic_glob = xstrdup (pattern);
-    d->msg.fn = cb;
-    d->msg.arg = arg;
-    if (zlist_push (h->reactor->dsp, d) < 0)
-        oom ();
-    rc = 0;
-done:
-    return rc;
-}
-
-int flux_msghandler_addvec (flux_t h, msghandler_t *handlers, int len,
-                            void *arg)
-{
-    int i;
-
-    for (i = 0; i < len; i++)
-        if (flux_msghandler_add (h, handlers[i].typemask, handlers[i].pattern,
-                                    handlers[i].cb, arg) < 0)
-            return -1;
-    return 0;
-}
-
-void flux_msghandler_remove (flux_t h, int typemask, const char *pattern)
-{
-    dispatch_t *d;
-
-    d = zlist_first (h->reactor->dsp);
-    while (d) {
-        if (d->type == DSP_TYPE_MSG && d->msg.match.typemask == typemask
-                             && !strcmp (d->msg.match.topic_glob, pattern)) {
-            zlist_remove (h->reactor->dsp, d);
-            dispatch_destroy (d);
-            if (reactor_empty (h->reactor))
-                flux_reactor_stop (h);
-            break;
-        }
-        d = zlist_next (h->reactor->dsp);
-    }
-}
-
-int flux_fdhandler_add (flux_t h, int fd, short events,
-                        FluxFdHandler cb, void *arg)
-{
-    dispatch_t *d;
-    int rc = -1;
-
-    if (fd < 0 || events == 0 || !cb) {
-        errno = EINVAL;
-        goto done;
-    }
-    if (!h->ops->reactor_fd_add) {
-        errno = ENOSYS;
-        goto done;
-    }
-    if (h->ops->reactor_fd_add (h->impl, fd, events) < 0)
-        goto done;
-
-    d = dispatch_create (DSP_TYPE_FD);
-    d->fd.fd = fd;
-    d->fd.events = events;
-    d->fd.fn = cb;
-    d->fd.arg = arg;
-    if (zlist_append (h->reactor->dsp, d) < 0)
-        oom ();
-    rc = 0;
-done:
-    return rc;
-}
-
-void flux_fdhandler_remove (flux_t h, int fd, short events)
-{
-    dispatch_t *d;
-
-    d = zlist_first (h->reactor->dsp);
-    while (d) {
-        if (d->type == DSP_TYPE_FD && d->fd.fd == fd
-                                   && d->fd.events == events) {
-            zlist_remove (h->reactor->dsp, d);
-            dispatch_destroy (d);
-            if (reactor_empty (h->reactor))
-                flux_reactor_stop (h);
-            break;
-        }
-        d = zlist_next (h->reactor->dsp);
-    }
-    if (h->ops->reactor_fd_remove)
-        h->ops->reactor_fd_remove (h->impl, fd, events);
-}
-
-int flux_zshandler_add (flux_t h, void *zs, short events,
-                        FluxZsHandler cb, void *arg)
-{
-    dispatch_t *d;
-
-    if (!h->ops->reactor_zs_add) {
-        errno = ENOSYS;
-        return -1;
-    }
-    if (h->ops->reactor_zs_add (h->impl, zs, events) < 0)
-        return -1;
-    d = dispatch_create (DSP_TYPE_ZS);
-    d->zs.zs = zs;
-    d->zs.events = events;
-    d->zs.fn = cb;
-    d->zs.arg = arg;
-    if (zlist_append (h->reactor->dsp, d) < 0)
-        oom ();
-    return 0;
-}
-
-void flux_zshandler_remove (flux_t h, void *zs, short events)
-{
-    dispatch_t *d;
-
-    d = zlist_first (h->reactor->dsp);
-    while (d) {
-        if (d->type == DSP_TYPE_ZS && d->zs.zs == zs
-                                   && d->zs.events == events) {
-            zlist_remove (h->reactor->dsp, d);
-            dispatch_destroy (d);
-            if (reactor_empty (h->reactor))
-                flux_reactor_stop (h);
-            break;
-        }
-        d = zlist_next (h->reactor->dsp);
-    }
-    if (h->ops->reactor_zs_remove)
-        h->ops->reactor_zs_remove (h->impl, zs, events);
-}
-
-int flux_tmouthandler_add (flux_t h, unsigned long msec, bool oneshot,
-                           FluxTmoutHandler cb, void *arg)
-{
-    dispatch_t *d;
-    int id;
-
-    if (!h->ops->reactor_tmout_add) {
-        errno = ENOSYS;
-        return -1;
-    }
-    if ((id = h->ops->reactor_tmout_add (h->impl, msec, oneshot)) < 0)
-        return -1;
-    d = dispatch_create (DSP_TYPE_TMOUT);
-    d->tmout.fn = cb;
-    d->tmout.arg = arg;
-    d->tmout.timer_id = id;
-    d->tmout.oneshot = oneshot;
-    if (zlist_append (h->reactor->dsp, d) < 0)
-        oom ();
-    return id;
-}
-
-void flux_tmouthandler_remove (flux_t h, int timer_id)
-{
-    dispatch_t *d;
-
-    d = zlist_first (h->reactor->dsp);
-    while (d) {
-        if (d->type == DSP_TYPE_TMOUT && d->tmout.timer_id == timer_id) {
-            zlist_remove (h->reactor->dsp, d);
-            dispatch_destroy (d);
-            if (reactor_empty (h->reactor))
-                flux_reactor_stop (h);
-            break;
-        }
-        d = zlist_next (h->reactor->dsp);
-    }
-    if (h->ops->reactor_tmout_remove)
-        h->ops->reactor_tmout_remove (h->impl, timer_id);
-}
-
-int flux_reactor_start (flux_t h)
-{
-    if (!h->ops->reactor_start) {
-        errno = ENOSYS;
-        return -1;
-    }
-    return h->ops->reactor_start (h->impl);
-}
-
-void flux_reactor_stop (flux_t h)
-{
-    if (h->ops->reactor_stop)
-        h->ops->reactor_stop (h->impl, 0);
+    return h->reactor;
 }
 
 /*
